+++ src/main/java/com/innowise/employeemodule/service/PositionServiceImpl.java
@@ -7,6 +7,7 @@ import com.innowise.employeemodule.repository.PositionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityExistsException;
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
@@ -29,17 +30,29 @@ public class PositionServiceImpl implements PositionService {
 
     @Override
     public Position add(Position position) {
+        if(repository.existsByName(position.getName())){
+            throw new EntityExistsException("Position with name: '" + position.getName() + "' already exists");
+        }
         position.setActive(true);
         return repository.save(position);
     }
 
     @Override
     public Position update(Position position) {
+        if(!repository.existsByName(position.getName())){
+            throw new EntityNotFoundException("Position with name: '" + position.getName() + "' not found");
+        }
+        if(!repository.existsById(position.getId())){
+            throw new EntityNotFoundException("Position with id: '" + position.getId() + "' not found");
+        }
         return repository.save(position);
     }
 
     @Override
     public void deleteById(Long id) {
+        if(!repository.existsById(id)){
+            throw new EntityNotFoundException("Position with id: '" + id + "' not found");
+        }
         repository.deleteById(id);
     }
